requirements.txt
# Core Libraries
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.4.2

# Music and MIDI processing
music21==8.1.0

# Machine Learning (Intel Optimized)
tensorflow-intel==2.13.0

# Metaheuristic Optimization
mealpy==1.0.2

# Optional utilities
tqdm==4.66.1
requests==2.31.0
